# docs/swagger.yaml
definitions:
...
  dto.BindUserRole:
    properties:
      roleIds:
        items:
          type: integer
        type: array
      userId:
        type: integer
    required:
    - roleIds
    - userId
    type: object
...
  dto.CreateUser:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  dto.DeleteUser:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  dto.DetailUser:
    properties:
      id:
        type: integer
    type: object
  dto.Login:
    properties:
      captcha:
        type: string
      captchaId:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captcha
    - captchaId
    - password
    - username
    type: object
...
  dto.SearchUser:
    properties:
      page:
        type: integer
      password:
        type: string
      size:
        type: integer
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  dto.UpdateUser:
    properties:
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
paths:
...
  /base/login:
    post:
      parameters:
      - description: 用户名, 密码, 验证码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.Login'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"登陆成功"}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - Base
...
  /user/all:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取所有的User 不分页
      tags:
      - User
  /user/bindRole:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户id, 角色id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.BindUserRole'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"绑定成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 给用户绑定角色权限
      tags:
      - User
  /user/create:
    post:
      parameters:
      - description: 用户名, 昵称, 密码, 角色ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"注册成功"}'
          schema:
            type: string
      summary: 创建用户
      tags:
      - User
  /user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除User
      tags:
      - User
  /user/detail:
    post:
      consumes:
      - application/json
      parameters:
      - description: 根据id获取User
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DetailUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取User
      tags:
      - User
  /user/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页获取User列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SearchUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取User列表
      tags:
      - User
  /user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新User
      tags:
      - User
swagger: "2.0"